backend/app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
openai==1.3.0
pytesseract==0.3.10
Pillow==10.1.0
PyPDF2==3.0.1
pdf2image==1.16.3
python-dotenv==1.0.0

